refactor(auth): move UI logic from AuthFeature to AuthService
The AuthFeature class was removed and its UI initialization and session handling logic was integrated into AuthService. This consolidates authentication responsibilities into a single service class, simplifying the feature registration process and improving code organization.

The AuthService now directly manages the authentication window and session updates, removing the need for a separate feature class. The init() method in Main.java now calls AuthService.getInstance().init() instead of registering AuthFeature.
